+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -25,6 +25,7 @@ from .test_create_tag import GmpCreateTagTestCase
 from .test_create_target import GmpCreateTargetCommandTestCase
 from .test_get_aggregates import GmpGetAggregatesTestCase
 from .test_get_info import GmpGetInfoTestCase
+from .test_get_info_list import GmpGetInfoListTestCase
 from .test_modify_filter import GmpModifyFilterTestCase
 from .test_modify_permission import GmpModifyPermissionTestCase
 from .test_modify_report_format import GmpModifyReportFormatTestCase